moved CNF parsing into separate file
